Added Condition Action, Hash table changed to dictionary

+++ TibiaHeleper/Modules/WalkerModule/WalkerLabel.cs
@@ -12,7 +12,7 @@ namespace TibiaHeleper.Modules.WalkerModule
         public WalkerLabel(string name)
         {
             this.name = name;
-            this.type = (int)StatementType.getType["label"];
+            this.type = StatementType.getType["label"];
         }
 
         public override object Clone()
